% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remuse.R
\docType{package}
\name{ReMUSE-package}
\alias{ReMUSE}
\alias{ReMUSE-package}
\title{Package for the classe "R e-MUSE"
Contains the slides as vignettes, the "fruits" dataset, the "nutrimenu" dataset,  and the tutorials. https://vguillemot.github.io/remuse/}
\description{
This package contains all the slides and data-sets used during the course ReMUSE.
}
\author{
\strong{Maintainer}: Vincent Guillemot \email{vincent.guillemot@pasteur.fr}

}
\keyword{internal}
